refactor: useDetailPost 분리
